+++ tests/integration/pipelex/pipes/controller/pipe_sequence/test_pipe_sequence_validation.py
@@ -19,12 +19,12 @@ class TestPipeSequenceValidation:
         concept_1 = ConceptFactory.make_from_blueprint(
             concept_code="TestConcept",
             domain=domain,
-            blueprint=ConceptBlueprint(description="Lorem Ipsum"),
+            blueprint_or_string_description=ConceptBlueprint(description="Lorem Ipsum"),
         )
         concept_2 = ConceptFactory.make_from_blueprint(
             concept_code="ProcessedText",
             domain=domain,
-            blueprint=ConceptBlueprint(description="Lorem Ipsum"),
+            blueprint_or_string_description=ConceptBlueprint(description="Lorem Ipsum"),
         )
         concept_library = get_concept_library()
         concept_library.add_concepts([concept_1, concept_2])
@@ -57,12 +57,12 @@ class TestPipeSequenceValidation:
         concept_1 = ConceptFactory.make_from_blueprint(
             concept_code="TestConcept",
             domain=domain,
-            blueprint=ConceptBlueprint(description="Lorem Ipsum"),
+            blueprint_or_string_description=ConceptBlueprint(description="Lorem Ipsum"),
         )
         concept_2 = ConceptFactory.make_from_blueprint(
             concept_code="ProcessedText",
             domain=domain,
-            blueprint=ConceptBlueprint(description="Lorem Ipsum"),
+            blueprint_or_string_description=ConceptBlueprint(description="Lorem Ipsum"),
         )
         concept_library = get_concept_library()
         concept_library.add_concepts([concept_1, concept_2])